Refactor tracking helper functions
